+++ packages/components/src/Calendar/Calendar.test.tsx
@@ -33,6 +33,8 @@ const theme: Theme = {
       calendarControllers: cva(),
       calendarHeader: cva(['fontWeight:bolder p-2']),
       calendarGrid: cva('[&_td]:p-2'),
+      select: cva('flex gap-1'),
+      calendarListboxButton: cva('bg-red-300'),
     },
     Select: cva() as any,
     ListBox: cva() as any,