+++ themes/theme-b2b/src/components/Calendar.styles.ts
@@ -14,4 +14,19 @@ export const Calendar: ThemeComponent<'Calendar'> = {
   ]),
   calendarHeader: cva(['fontWeight:bolder p-2']),
   calendarGrid: cva('[&_td]:p-2'),
+  calendarListboxButton: cva([
+    'rounded-xs border-none px-4 leading-[48px] h-component outline-hidden',
+    'focus-visible:outline-outline-focus focus-visible:outline focus-visible:outline-offset-1',
+    'hover:bg-bg-brand-hover hover:text-text-inverted',
+    'aria-[current=true]:text-text-inverted aria-[current=true]:bg-bg-brand',
+    'aria-[current=true]:hover:bg-bg-brand-hover',
+  ]),
+  select: cva([
+    'bg-bg-surface',
+    'px-2',
+    'border-border-base rounded-xs border outline-hidden outline-offset-[-1px] h-component',
+    'focus:outline-outline-focus focus:outline-2',
+    'hover:border-border-base-hover',
+    'disabled:bg-bg-base-disabled disabled:text-text-base-disabled disabled:hover:border-border-base-disabled disabled:border-border-base-disabled disabled:cursor-not-allowed',
+  ]),
 };