+++ themes/theme-core/src/components/Calendar.styles.ts
@@ -12,4 +12,15 @@ export const Calendar: ThemeComponent<'Calendar'> = {
   calendarControllers: cva(),
   calendarHeader: cva(['fontWeight:bolder p-2']),
   calendarGrid: cva('[&_td]:p-2'),
+  calendarListboxButton: cva([
+    'text-text-base ease-ease-out h-component cursor-pointer rounded-xs px-4 py-0 text-sm leading-[22px] transition-all duration-200',
+    'focus-visible:outline-outline-focus focus-visible:outline focus-visible:outline-offset-1 outline-hidden',
+    'hover:bg-bg-inverted-hover',
+    'aria-[current=true]:border-border-brand aria-[current=true]:bg-bg-brand aria-[current=true]:text-text-inverted aria-[current=true]:hover:bg-bg-brand-hover aria-[current=true]:hover:border-border-brand-hover',
+  ]),
+  select: cva([
+    'h-component px-1',
+    'border rounded-xs border-border-inverted bg-white',
+    'disabled:bg-bg-inverted-disabled disabled:border-border-base-disabled disabled:text-text-base-disabled disabled:cursor-not-allowed',
+  ]),
 };